moving report and benchmarks folder
