+++ test/benchmarks/console-log.js
